Corrección de eliinación de proveedor

+++ app/Http/Controllers/ProviderController.php
@@ -3,6 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Provider;
+use App\Capture;
+use App\Statement;
+use App\Product;
 use App\Address;
 use DB;
 use Illuminate\Http\Request;
@@ -29,6 +32,7 @@ class ProviderController extends Controller
       $providers = DB::table('providers')
         ->select('providers.*', 'providers.id as provider_id', 'addresses.*', 'addresses.id as address_id')
         ->join('addresses', 'addresses.provider_id', '=', 'providers.id')
+        ->where('providers.status', '=', 1)
         ->get();
 
         for($i=0; $i<$providers->count(); $i++)
@@ -152,13 +156,26 @@ class ProviderController extends Controller
      */
     public function destroy(Provider $provider)
     {
-        $provider->delete();
+        // 0 = muerto, 1 = vivo
+        $captures = Capture::where('provider_id', $provider->id)->first();
+        $products = Product::where('provider_id', $provider->id)->first();
+        $statement = Statement::where('provider_id', $provider->id)->first();
+        if($captures == null && $products == null)
+        {
+            $provider->delete();
+        }
+        else
+        {
+            $provider->status = 0;
+            $provider->save();
+            $statement->status=0;
+            $statement->save();
+        }
         $msg = [
             'title' => 'Eliminado!',
             'text' => 'Proveedor eliminado exitosamente.',
             'icon' => 'success'
         ];
-
         return response()->json($msg);
     }
 }
